use_webhooks/backend/python/pyproject.toml
[tool.poetry]
name = "create_link_app"
version = "0.1.0"
description = ""
authors = ["moneykit.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
moneykit = "^0.1.4"
fastapi = "^0.104.1"
uvicorn = "^0.24.0.post1"
pydantic-settings = "^2.1.0"
httpx = "^0.25.2"
cachetools = "^5.3.2"

pyjwt = { version = "^2.8.0", extras = ["crypto"] }

[tool.poetry.group.dev.dependencies]
watchgod = "^0.8.2"
ruff = "^0.1.6"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py311"]

[tool.ruff]
select = ["I", "E", "W", "F", "C"]
ignore = ["E501"]
line-length = 120

[tool.mypy]
python_version = "3.11"
plugins = ["pydantic.mypy"]
exclude = [".venv/"]
warn_return_any = false
warn_redundant_casts = true
warn_unused_ignores = true
warn_unused_configs = true
disallow_untyped_calls = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_optional = true
ignore_missing_imports = true
